Add files via upload
go to directory and use the same function "python main.py acceleration_filename.csv" in cmd.
The code reads a csv file as a pandas dataframe, plots the raw data, parses it through a butterworth filter (questionable, but easy to apply and can be changed later on), and then plots the filtered data. Everything is done in dataframes.